+++ component-classes/index.d.ts
@@ -491,4 +491,20 @@ export namespace backwardsCompatibleClasses {
     export { radioInvalid_1 as radioInvalid };
     const radioDisabled_1: string;
     export { radioDisabled_1 as radioDisabled };
+    export const modalTitle: string;
+    export const pillClose: string;
+}
+export namespace pagination {
+    export const currentPage: string;
+    const icon_4: string;
+    export { icon_4 as icon };
+    export const containerNav: string;
+    export const heading: string;
+    export const pages: string;
+    export const nextPage: string;
+    export const firstPageButton: string;
+    export const firstPageLabel: string;
+    const active_1: string;
+    export { active_1 as active };
+    export const defaultPage: string;
 }